Fix AMP dtype mismatch at ground truth loading stage
Problem:
- Previous fix converted targets after assignment, but dtype mismatch
  occurred during the assignment itself (lines 865-867)
- Ground truth tensors loaded as Float32 but assigned to Float16 targets
- Error: "expected scalar type Float but found Half"

Root Cause:
- gt_depth_flat, gt_dims_flat, gt_rot_flat loaded as Float32
- target_depth, target_dims, target_rot created with torch.zeros_like(params_3d)
  which are Float16 with AMP enabled
- Assignment operation fails due to dtype mismatch

Solution:
- Convert ground truth tensors to params_3d.dtype at load time
- Add dtype parameter to .to() call: .to(self.device, dtype=params_3d.dtype)
- Ensures all tensors have consistent dtype from the start

Changes in ultralytics/utils/loss.py (lines 830-832):
  gt_depth_flat = batch["location_3d"].to(self.device, dtype=params_3d.dtype)[:, 2:3]
  gt_dims_flat = batch["dimensions_3d"].to(self.device, dtype=params_3d.dtype)
  gt_rot_flat = batch["rotation_y"].to(self.device, dtype=params_3d.dtype)

This fully enables GPU training with AMP for YOLOv12-3D.

+++ ultralytics/utils/loss.py
@@ -826,9 +826,10 @@ class v8Detection3DLoss(v8DetectionLoss):
                 # location_3d shape: [N_total, 3] -> [x, y, z] (we mainly care about z for depth)
                 # rotation_y shape: [N_total, 1]
                 batch_idx_gt = batch["batch_idx"].view(-1).long().to(self.device)
-                gt_depth_flat = batch["location_3d"].to(self.device)[:, 2:3]  # z coordinate
-                gt_dims_flat = batch["dimensions_3d"].to(self.device)  # [h,w,l]
-                gt_rot_flat = batch["rotation_y"].to(self.device)
+                # Convert to same dtype as params_3d for AMP compatibility
+                gt_depth_flat = batch["location_3d"].to(self.device, dtype=params_3d.dtype)[:, 2:3]  # z coordinate
+                gt_dims_flat = batch["dimensions_3d"].to(self.device, dtype=params_3d.dtype)  # [h,w,l]
+                gt_rot_flat = batch["rotation_y"].to(self.device, dtype=params_3d.dtype)
 
                 # Assign 3D targets to positive anchors
                 target_depth = torch.zeros_like(params_3d[:, :, 0:1])
